guided-projects/managed-mcp-bigquery-maps-agent/requirements.txt
pandas
numpy
google-cloud-bigquery
google-adk
python-dotenv
